### 2023-05-31 #5509 @editorialbot
[PRE REVIEW]: Time-frequency component of the Green-X library: minimax grids for efficient RPA and GW calculations
**Submitting author:** @aziziph (<a href="http://orcid.org/0000-0001-9089-1043">Maryam Azizi</a>)
**Repository:** https://github.com/nomad-coe/greenX
**Branch with paper.md** (empty if default branch): main
**Version:** V1.0.0
