+++ JOSS/refs.bib
@@ -6,6 +6,7 @@
   number={8},
   pages={5140--5154},
   year={2021},
+  doi = {10.1021/acs.jctc.1c00488},
   publisher={ACS Publications}
 }
 @article{ren2015beyond,
@@ -16,6 +17,7 @@
   number={8},
   pages={081104},
   year={2015},
+  doi= {10.1103/PhysRevB.92.081104},
   publisher={APS}
 }
 @article{Pearson:2017,
@@ -26,6 +28,7 @@
   	Journal = {ArXiv e-prints},
   	Keywords = {Astrophysics - Astrophysics of Galaxies},
   	Month = mar,
+    doi = {10.1038/s41550-017-0220-3},
   	Title = {{Gaps in Globular Cluster Streams: Pal 5 and the Galactic Bar}},
   	Year = 2017
 }
@@ -34,6 +37,7 @@
   	Author = {{Azizi}, M. and {Wilhelm}, J. and {Panadés-Barrueta}, R.~L. and {Golze}, D. and {Giantomassi}, M. and {Delesma}, F.~A. and {Rinke}, P. and {Gonze}, X.},
   	Journal = {Unpublished},
     year = {2023},
+    doi = {--},
   	Title = {{Validation of the Green-X library time-frequency component for GW and RPA calculations}},
 }
 @book{Binney:2008,
@@ -42,6 +46,7 @@
   	Booktitle = {Galactic Dynamics: Second Edition, by James Binney and Scott Tremaine.~ISBN 978-0-691-13026-2 (HB).~Published by Princeton University Press, Princeton, NJ USA, 2008.},
   	Publisher = {Princeton University Press},
   	Title = {{Galactic Dynamics: Second Edition}},
+    doi = {10.2307/j.ctvc778ff},
   	Year = 2008
 }
 
@@ -81,6 +86,7 @@
   year = {2020},
   publisher = {GitHub},
   journal = {GitHub repository},
+  doi = {https://doi.org/10.6084/m9.figshare.828487},
   url = {https://github.com/arfon/fidgit}
 }
 
@@ -91,6 +97,7 @@
   author={Braess, Dietrich},
   volume={7},
   year={2012},
+  doi = {10.1007/978-3-642-61609-9},
   publisher={Springer Science \& Business Media}
 }
 @article{golze2019gw,
@@ -100,6 +107,7 @@
   volume={7},
   pages={377},
   year={2019},
+  doi = {10.3389/fchem.2019.00377},
   publisher={Frontiers}
 }
 @article{bruneval2008accurate,
@@ -110,12 +118,14 @@
   number={8},
   pages={085125},
   year={2008},
+  doi = {10.1103/PhysRevB.78.085125 },
   publisher={APS}
 }
 @book{press2007numerical,
   title={Numerical recipes 3rd edition: The art of scientific computing},
   author={Press, William H and Teukolsky, Saul A and Vetterling, William T and Flannery, Brian P},
   year={2007},
+  doi = {10.1145/1874391.187410},
   publisher={Cambridge university press}
 }
 @article{hackbusch2019computation,
@@ -126,6 +136,7 @@
   number={1-2},
   pages={1--11},
   year={2019},
+  doi = {10.1007/s00791-018-00308-4},
   publisher={Springer}
 }
 @article{Takatsuka2008,
@@ -136,6 +147,7 @@
     number = {4},
     year = {2008},
     month = {07},
+    doi = {10.1007/s00791-018-00308-4}
 }
 @article{Adler1962,
     author = "Adler, Stephen L.",
@@ -147,6 +159,7 @@
[…]
