+++ JOSS/refs.bib
@@ -726,7 +726,7 @@ year = {2013},
   number={12},
   pages={5851--5859},
   year={2016},
-  doi = {10.1021/acs.jctc.6b00840}
+  doi = {10.1021/acs.jctc.6b00840},
   publisher={ACS Publications}
 }
 @article{wilhelm2016rpa,
