+++ JOSS/refs.bib
@@ -288,6 +288,7 @@
   number={24},
   pages={241201},
   year={2011},
+  doi = {10.1103/PhysRevB.84.241201},
   publisher={APS}
 }
 @article{li2005quasiparticle,
@@ -298,6 +299,7 @@
   number={19},
   pages={193102},
   year={2005},
+  doi = {10.1103/PhysRevB.71.193102},
   publisher={APS}
 }
 @article{wilhelm2016gw,
@@ -319,6 +321,7 @@
   number={1},
   pages={1--11},
   year={2020},
+  doi = {10.1038/s41597-020-0385-y},
   publisher={Nature Publishing Group}
 }
 @article{forster2020low,
@@ -362,6 +365,7 @@
   number={5},
   pages={053020},
   year={2012},
+  doi = {10.1088/1367-2630/14/5/053020},
   publisher={IOP Publishing}
 }
 @article{van2013gw,
@@ -562,6 +566,7 @@ YEAR={2021},
       author={Arno Förster and Erik van Lenthe and Edoardo Spadetto and Lucas Visscher},
       year={2023},
       pages={2303.09979},
+      doi = {10.48550/arXiv.2303.09979},
       journal={arXiv},
 }
 
@@ -645,6 +650,7 @@ year = {2004},
   number={3},
   pages={035118},
   year={2016},
+  doi = {10.1103/PhysRevB.94.035118},
   publisher={APS}
 }
 @article{ren2012random,
@@ -736,6 +742,7 @@ journal = {Journal of Chemical Theory and Computation},
 volume = {12},
 number = {12},
 pages = {5851-5859},
+doi = {10.1021/acs.jctc.6b00840},
 year = {2016},
 }
 @article{kutepov2012electronic,
@@ -746,6 +753,7 @@ year = {2016},
   number={15},
   pages={155129},
   year={2012},
+  doi = {10.1103/PhysRevB.85.155129},
   publisher={APS}
 }
 @article{kutepov2017linearized,
