### 2023-06-01 #5509 @lucydot
Hi @aziziph, thank you for your submission to JOSS.

I'll start looking for reviewers. If you could do the following it will help move things along:

- if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.
- please fix the missing doi's list above ☝️
- JOSS papers are short advertisements for the software package, and as a guideline the word count is 250-1000 words. Your paper currently exceeds this guideline by a significant amount. I suggest that the `mathematical framework` section is moved online; you can then link to this from the paper.
