### 2023-06-05 #5509 @lucydot
@aziziph I can see that the JOSS editorial bot and github have not automatically detected the license file. I suggest renaming this to the more standard `LICENSE.txt` so that it is automatically detected.
