### 2023-06-05 #5509 @lucydot
@aziziph please also take a look at the three invalid DOI's listed above; you can check yourself if they compile correctly using `@editorialbot check references`.
