+++ JOSS/refs.bib
@@ -86,7 +86,7 @@
   year = {2020},
   publisher = {GitHub},
   journal = {GitHub repository},
-  doi = {https://doi.org/10.6084/m9.figshare.828487},
+  doi = {10.6084/m9.figshare.828487},
   url = {https://github.com/arfon/fidgit}
 }
 
