### 2023-06-18 #5509 @lucydot
OK, thanks for agreeing to review @mailhexu and @DarioALeonValido ⭐

ping @aziziph

I'll ask @editorialbot to start the review proper (which will be on a separate thread). There will be more instructions there - the first reviewer task will be to generate a review checklist to work through. The process is different from traditional journals - it is an on-going conversation between reviewers, authors, editors and our bot. So I recommend starting the review early and keeping us updated as things progress rather than trying to complete the review or response to a review in a single-shot. If you have any questions, you can raise them on the review thread.
