### 2023-06-21 #5570 @DarioALeonValido
@aziziph

Dear Maryam and other authors, I find this library potentially very useful. However, my main concern so far is regarding validation and performance, that are left to an unpublished paper (Azizi et al 2023).

I suggest you to include here at lest a simple example illustrating the efficiency of the minimax grids. Maybe it could be done by integrating a toy function, but I also would like to see if you can get an estimate of how much it is reduced the prefactor of low-scaling MP2/RPA/GW calculations to justify the exclusive use of this particular quadrature.

Another related question. In line 150  of the pdf you specify that the number of time/frequency points in the library ranges from 2 to 64, how are these values related to the final accuracy of correlation and quasi-particle energies?
