### 2023-06-21 14c10bf
Update minimax_grids.F90
