### 2023-06-28 #5570 @lucydot
Hi @aziziph, re-pinging the message above.

The review process in JOSS is different from other journals in that it is more of a conversation back and forth (rather than a review that is completed in one go, which is then responded to in one go). My advice would be to start responding to points as they are raised. Usually the review process will include several merged PRs, and the reviewers keep track using the checklist.

As mentioned in the pre-review, I suggest renaming the license file to LICENSE.txt as that is standard for gh repos (and will be automatically recognised as a license by gh), and there are a couple of invalid DOIs reported in the thread above.
