### 2023-07-02 nomad-coe/greenX#48 @mailhexu
Information to be added in License file
In the License file, the following line from the template of the Apache 2 licence should be replaced.
  Copyright [yyyy] [name of copyright owner]
