### 2023-07-02 nomad-coe/greenX#49 @mailhexu
Documentation compilation disabled by default
In the README.md, "GreenX is documented using Doxygen, and documentation support is enabled by default."
But in the cmake file it seems to be disabled by default. I could only build the documentations by manually switching it to ON.
