### 2023-07-02 a1bc51e
Update README.md: command for installing zofu
cd zofu 
is missing after cloning zofu form github.
