### 2023-07-02 nomad-coe/greenX#51 @mailhexu
requirement of pygreenx  for unittest   should be added to README.md
It is not clear from the documentation that pygreenx should be installed for the unit tests.  And the way to install it should be added too, although the source code can be found in the python directory. Indeed without knowing that the unit tests 2 and 3 will fail.
