### 2023-07-02 #5570 @lucydot
Hello @mailhexu @aziziph @DarioALeonValido
Just to let you know that I am at a conference for the next week; will check in again w/c 10th July.
