### 2023-07-02 #5570 @mailhexu
@aziziph  Dear Mariam and other authors,

Thanks for making this library available! I think it can be very useful.

I found a few minor issues during the reviewing procedure:
I opened a few minor issues:
https://github.com/nomad-coe/greenX/issues/48
https://github.com/nomad-coe/greenX/issues/49
https://github.com/nomad-coe/greenX/issues/51
https://github.com/nomad-coe/greenX/issues/52
And a PR:
https://github.com/nomad-coe/greenX/pull/50

I have some minor suggestions below:
- I find it unclear whether some part of  the Summary is referring to GreenX as a whole or specifically discussing its time-frequency component. For instance, in line 22, "the package" appears to refer to the GreenX library, while the subsequent sentences seem to focus more on the time-frequency component. Using an acronym like GXTF for the component could potentially provide clarity even though the time-frequency library is currently the only functional component in GreenX, based on the presence of multiple directories with the prefix "GreenX-" with only "README" files inside the main branch (correct me if the other components are already developped). This might confuse the fut
- I didn't find the mentioning of the documentation of the API in the manuscript. The method of compiling and opening  the documentation is indeed in the README.md, but   it doesn't seem straightforward to know what is inside. Some introduction of the contents should be helpful. To put it online (e.g. in github pages or readthedocs.io) can also be helpful.
- An example code of using it (as provided in the doxygen documentation) is helpful if it is made more easily accessible (e.g. in an example directory).
