### 2023-07-04 nomad-coe/greenX#51 @AlexBuccheri
Could @mailhexu be more specific as to what they'd like to see, and where?

In the top-level README, it states:

```
## Running the Tests

GreenX uses pytest as its regression testing framework, in conjunction with
some custom python modules. First, one must ensure that the python utilities
are installed. From the GreenX root directory:

```bash
cd python
pip install -e .

In the time-frequency README.md, it states:

```
## Running the Tests

Application tests are run with the pytest framework. Having installed `pygreenx`
(see the top-level README) change to `<GX_ROOT>/<BUILD_DIR>` and type `ctest`.
Additionally, one can change to the test folder and explicitly run the pytest
command from there:

```bash
cd <GX_ROOT>/<BUILD_DIR>/test/time-frequency
pytest -s
# or
pytest -s test_name.py

Things I could think to improve this are:

- Explicitly mention the name pygreenx in the top-level details (instead of saying python utilities)
- Have CMake check if pygreenx is installed, and give instructions if not
