### 2023-07-04 nomad-coe/greenX#49 @AlexBuccheri
If documentation were on by default, the  lack of an available doxygen package on a given system would break the build. Doxygen should be an optional dependency.

@AlexBuccheri

- [ ] Update documentation stating that documentation is off by default.
