### 2023-07-04 nomad-coe/greenX#52 @AlexBuccheri
This is the last element of the array, with a diff of 3.2e-7. Just increase the tolerance slightly:

- [ ] Change atol to 5.e-7
