### 2023-07-04 nomad-coe/greenX#51 @mailhexu
Sorry that I didn't notice the first part.  When I was trying the unit test part, I only searched in the Unit Test section.
But indeed what you proposed for the improvement is nice.
