### 2023-07-04 5918e8d
Update test_gx_tabulate_minimax.py change atol=1.e-7 to atol=5.e-7 for large grids
