### 2023-07-04 nomad-coe/greenX#52 @fdelesma
Could you review the request #53 ?
