### 2023-07-04 nomad-coe/greenX#51 @AlexBuccheri
@mailhexu In that case I'll reopen the issue such that I can create an MR that addresses the points I've suggested
