+++ python/pygreenx/__init__.py
@@ -1,0 +1,3 @@
+from pkg_resources import get_distribution
+
+__version__ = get_distribution('pygreenx').version
