### 2023-07-05 018cc18
Change top-level CMakeLists.txt to only search for pygreenx, and inform the user if it is not found. Move responsibility of other python packages to pygreenx. Python version now explicit in top-level CMakeLists.txt.
Change function to find python modules such that it optionally accepts REQUIRED and VERSION.
