### 2023-07-05 685befd
Correct README to be consistent with CMake default settings for documentation build.
