### 2023-07-05 5b8b40e
Return variable ${module}_FOUND, and correct typo.
