### 2023-07-05 #5570 @AlexBuccheri
@mailhexu Not sure it will let me tag you as a reviewer, but here's the MR addressing issues 49 and 51:

https://github.com/nomad-coe/greenX/pull/54
