+++ GX-TimeFrequency/test/test_gx_tabulate_minimax.py
@@ -111,11 +111,11 @@ def test_tabulate_gx_minimax_grid(fortran_binary):
     # Grids for which values do get large or very small
     assert np.allclose(tabulated_errors_large_grids[:, Column.MaxErrCosFTTimeToFreq.value],
                        ref_errors_big_grid[:, Column.MaxErrCosFTTimeToFreq.value],
-                       atol=1.e-7)
+                       atol=1.e-6)
 
     assert np.allclose(tabulated_errors_large_grids[:, Column.MaxErrCosFTFreqToTime.value],
                        ref_errors_big_grid[:, Column.MaxErrCosFTFreqToTime.value],
-                       atol=1.e-7)
+                       atol=1.e-6)
 
     assert np.allclose(tabulated_errors_large_grids[:, Column.MaxErrSinFTimeToFreq.value],
                        ref_errors_big_grid[:, Column.MaxErrSinFTimeToFreq.value],
