+++ CMakeLists.txt
@@ -44,11 +44,24 @@ find_package(LAPACK REQUIRED)
 
 # Python required for application testing
 include(cmake/python3.cmake)
-find_python_module(numpy)
-find_python_module(pytest)
+
+find_package(Python3 3.7 COMPONENTS Interpreter Development)
+if(Python3_FOUND)
+    message("-- Python 3 interpreter version: " ${Python3_VERSION})
+else()
+    message("-- Python 3 interpreter not found")
+endif()
 
 # Enable ctest
 enable_testing()
+
+# Application test python module
+# Note, pygreenx dependencies should be handled pygreenx
+# pygreenx version specified in its setup.cfg
+find_python_module(pygreenx VERSION 0.0.1)
+if (NOT pygreenx_FOUND)
+    message("-- pygreenx is required for application testing. To install it, cd <PROJECTROOT>/python and run `pip install -e .`")
+endif ()
 
 # Optional unit testing lib
 option(ENABLE_UNITTESTS "Enable GreenX Unit Testing" OFF)
@@ -66,7 +79,7 @@ add_subdirectory(GX-common)
 add_subdirectory(GX-TimeFrequency)
 
 # Documentation
-option(ENABLE_DOCS "Enable documentation" ON)
+option(ENABLE_DOCS "Enable documentation" OFF)
 if (${ENABLE_DOCS})
     find_program(DOXYGEN doxygen REQUIRED)
     message("-- Doxygen documentation support enabled.")
