### 2023-07-10 nomad-coe/greenX#59 @Panadestein
Add short description, tags, and possibly website to the repo
Hello, as the title suggest, it would be nice to add a short sentence summarizing what Green-X is, and also tags to the GitHub repository. We could also link the NOMAD website there (or ideally deploy a GitHub pages there). See the image below, this is empty in our repo:

![image](https://github.com/nomad-coe/greenX/assets/15854703/99cecb72-bc0e-4080-9c70-03ce35c709b2)

This is a very common thing, and it helps to project gain visibility. Unfortunately, I do not have permission to do it myself. Can you take care of it @aziziph or @AlexBuccheri ?
