### 2023-07-10 nomad-coe/greenX#59 @aziziph
Hello, Do you propose this structure as a WIKI for our repository?
