### 2023-07-10 nomad-coe/greenX#59 @Panadestein
No, for the moment only adding a description. If you have the permissions, you should be able to directly edit the `About` section on GitHub.
