### 2023-07-10 nomad-coe/greenX#59 @AlexBuccheri
Yep, I have owner rights so I could assign repo settings.  Do you have a suggestion regarding the blurb?

>  link the NOMAD website there

No problem

> or ideally deploy a GitHub pages there

Could you clarify what you mean?
