### 2023-07-10 nomad-coe/greenX#59 @Panadestein
Oh, one could do a per-project GitHub [pages](https://pages.github.com/), but it does not have to be, it is only an idea. These pages are easy to set, and there are a bunch of Jekyll templates to choose from.
