### 2023-07-12 #5570 @aziziph
> @aziziph
>
> Dear Maryam and other authors, I find this library potentially very useful. However, my main concern so far is regarding validation and performance, that are left to an unpublished paper (Azizi et al 2023).
>
> I suggest you to include here at lest a simple example illustrating the efficiency of the minimax grids. Maybe it could be done by integrating a toy function, but I also would like to see if you can get an estimate of how much it is reduced the prefactor of low-scaling MP2/RPA/GW calculations to justify the exclusive use of this particular quadrature.
>
> Another related question. In line 150 of the pdf you specify that the number of time/frequency points in the library ranges from 2 to 64, how are these values related to the final accuracy of correlation and quasi-particle energies?

We thank the referee for the suggestion. The editor already remarked on the length of the paper "JOSS papers are short advertisements for the software package, and as a guideline the word count is 250-1000 words. Your paper currently exceeds this guideline by a significant amount." Since it is not possible to shorten the paper without sacrificing clarity, it seems that adding further material to the paper is not appropriate. However, we agree that it is highly interesting to compare minimax grids against standard time and frequency grids, which we will report extensively in an unpublished work, referenced as Azizi2023. We are happy to provide here an example. In this test, we evaluate the RPA total energy of CH4 using a Gauss-Legendre grid, a modified Gauss-Legendre grid (so far the standard in FHI-aims and abinit), and minimax grids. An accuracy of 10^-6 eV is reached with 10 minimax grid points while the modified Gauss-Legendre grids requires 36 points for this accuracy (Please see the following figure).
![CH4](https://github.com/openjournals/joss-reviews/assets/51956288/9c81c35a-4d94-40d1-a7ed-c92fbf493e80)
Error differences of the total RPA energy [eV] of methane calculated using the Gauss-Legendre, modified Gauss-Legendre and minimax imaginary frequency grid points. These differences were calculated with respect to the lowest RPA energy obtained with 34 minimax grid points. The ground state energy was calculated using the PBE exchange correlation functional in combination of the Tier2 basis set. The global resolution of identity (RI-V) approach was used for the calculation of the exact exchange and RPA correlation energy. The auxiliary basis functions for the RI-V method were generated automatically on the fly. #
