### 2023-07-13 #5570 @DarioALeonValido
@aziziph

Dear Maryam, thanks for providing this example. These results seem very promising indeed. Since you are using the minimax grid as a reference, I  wonder if the Gauss-Legendre grid is converging to a different value around 10^-3 eV. I guess you will comment on this in the benchmarking paper.

I see that all the issues raised by the other referee have been corrected, so I am happy to recommend the paper for publication without further ado.

All the best,
Dario A. Leon
