### 2023-07-13 #5570 @lucydot
Hi @DarioALeonValido - thanks for your expertise in reviewing the paper.

Hi @aziziph @AlexBuccheri - I agree that there is not space for the efficiency comparison in the JOSS paper. As this is the key performance claim of the software, a link to this comparison in the documentation or in a paper published elsewhere is required. Ideally it would be both. I note that the paper you link to is currently unpublished - is there a pre-print available?

I agree with @mailhexu that the documentation is quite inaccessible in its current form; requiring user installation to access (I think, it is not clear) function-level documentation. It is not easy for potential users to understand the software capabilities and use cases without reading the JOSS paper, which requires pdf download.

I suggest you make this information (software capabilities, use cases, API docs) available as webpages. This could be hosted on github pages, for example. At a minimum, this information (minus the API docs) need to be available on the README.md.

Happy to discuss this more if anything is unclear. Exploring previous repos published in JOSS can be a good way to find strong documentation examples. I can see this is being discussed on the issue [here](https://github.com/nomad-coe/greenX/issues/59).
