### 2023-07-13 #5570 @lucydot
A follow on note to say that by having a documentation webpage you will be able to cut the details in your JOSS paper. The JOSS paper is an advert for the repo, it does not need to go into details of the implementation (it is not the place for the e.g. mathematical framework section). This is better in the online docs.
