### 2023-07-13 nomad-coe/greenX#59 @Panadestein
Judging from the last comment from our editor @lucydot, this issue just gained in relevance. @AlexBuccheri @aziziph @fdelesma @dgolze please let's discuss this further.
