### 2023-07-13 #5570 @DarioALeonValido
@lucydot

Dear Lucy, thank you very much for providing this possible solution. I coincide with you that an illustrative example in support of the claims of the library should be available at a preprint of the benchmarking reference or somewhere else beside this thread.
