### 2023-07-20 #5570 @lucydot
Dear @aziziph @AlexBuccheri @panadestein  -

A heads up that I am going on annual leave until the 8th of August.

The primary outstanding issue relates to documentation, which I can see is under discussion (https://github.com/nomad-coe/greenX/issues/59). If you have any questions relating to this I will briefly have access to internet this Monday 24th, otherwise I will be off-grid!

There are also some more minor issues raised by @mailhexu (https://github.com/nomad-coe/greenX/issues/48 https://github.com/nomad-coe/greenX/issues/49 https://github.com/nomad-coe/greenX/issues/51 https://github.com/nomad-coe/greenX/issues/52).

Hopefully this gives a good timeline for completion; I will check back here ASAP after returning from AL as I think we are getting close to the end!

Lucy
