### 2023-07-24 ddf16e4
add https://arxiv.org/abs/2306.16066 to refs.bib
