+++ JOSS/refs.bib
@@ -766,7 +766,16 @@ year = {2016},
   doi = {10.1016/j.cpc.2017.06.012},
   publisher={Elsevier}
 }
-
+@article{Graml2023,
+  	Archiveprefix = {arXiv},
+    author={Maximilian Graml and Klaus Zollner and Daniel Hernangómez-Pérez and Faria Junior, Paulo E.  and Jan Wilhelm},
+    eprint={2306.16066},
+  	Journal = {ArXiv e-prints},
+  	Month = jul,
+    doi = {10.48550/arXiv.2306.16066},
+  	Title = {{Low-scaling GW algorithm applied to twisted transition-metal dichalcogenide heterobilayers}},
+  	Year = 2023
+}
 
 
 
