### 2023-07-24 bd342e9
Remove duplicate entries and fix erroneous unicode
