### 2023-07-24 a2c17b9
Format table, such that it's readable whilst editing.
Reformat the directory structure block, as unicode symbols not parsed by latex.
