### 2023-07-24 #5570 @AlexBuccheri
Hi @lucydot, I've opened an MR to address the documentation concerns. After a discussion, we collectively decided on putting the benchmarks on Github pages. This will still reside in the repo, but not the JOSS paper. @DarioALeonValido and @mailhexu can follow the progress [here](https://github.com/nomad-coe/greenX/pull/64)
