### 2023-07-24 nomad-coe/greenX#48 @AlexBuccheri
Resolved in MR #57
