### 2023-07-24 #5570 @AlexBuccheri
Wr.t. the open issues, they were actually all addressed. I'm not sure why Github didn't automatically close them (in the instances they were linked in the MRs)
