### 2023-07-25 bac20cb
Improved structure of gh pages
