### 2023-07-26 cffc651
Remove placeholders and apply formatting
