+++ docs/index.md
@@ -7,17 +7,9 @@ description: Library for Many-body Greens Functions on HPC
 
 # GreenX
 
-With [GitHub Pages](https://pages.github.com), you just write things in
-[Markdown](https://daringfireball.net/projects/markdown/),
-[GitHub](https://github.com) hosts the site for you, and you just push
-material to your GitHub repository with `git add`, `git commit`, and
-`git push`.
-
-The sites use [Jekyll](https://jekyllrb.com/), a
-[ruby](https://www.ruby-lang.org/en/) [gem](https://rubygems.org/), to
-convert Markdown files to html, and this part is done
-automatically when you push the materials to the `gh-pages` branch
-of a GitHub repository.
+A new open-source exascale library of Green-function-based methodologies. Its layered design will separate higher-level 
+functionalities (distinguishing between code-independent and code-family-specific parts) from architecture-dependent 
+numerical routines, common to all code families.
 
 ## Components
 
