### 2023-07-26 d00b256
Joss Responses: Library Documentation and Basic Profiling (#64)
* README documentation enhancement
* Addition of Github pages
* Removal of unicode characters from the JOSS md, and inclusion of local build instructions
---------

Co-authored-by: Panadestein <rpana92@gmail.com>
