+++ docs/index.md
@@ -1,0 +1,18 @@
+---
+layout: page
+title: GreenX Library
+tagline: GreenX
+description: Library for Many-body Greens Functions on HPC
+---
+
+# GreenX
+
+A new open-source exascale library of Green-function-based methodologies. Its layered design will separate higher-level 
+functionalities (distinguishing between code-independent and code-family-specific parts) from architecture-dependent 
+numerical routines, common to all code families.
+
+## Components
+
+- [Minimax Time-Frequency](gx_time_frequency.md)
+
+---
