+++ JOSS/refs.bib
@@ -1,83 +1,61 @@
-@article{wang2021assessing,
-  title={Assessing the G 0 W 0$\Gamma$0 (1) Approach: Beyond G 0 W 0 with Hedin’s Full Second-Order Self-Energy Contribution},
-  author={Wang, Yanyong and Rinke, Patrick and Ren, Xinguo},
-  journal={Journal of Chemical Theory and Computation},
-  volume={17},
-  number={8},
-  pages={5140--5154},
-  year={2021},
-  doi = {10.1021/acs.jctc.1c00488},
-  publisher={ACS Publications}
-}
-@article{ren2015beyond,
-  title={Beyond the G W approximation: A second-order screened exchange correction},
-  author={Ren, Xinguo and Marom, Noa and Caruso, Fabio and Scheffler, Matthias and Rinke, Patrick},
-  journal={Physical Review B},
-  volume={92},
-  number={8},
-  pages={081104},
-  year={2015},
-  doi= {10.1103/PhysRevB.92.081104},
-  publisher={APS}
-}
 @article{Pearson:2017,
-  	url = {http://adsabs.harvard.edu/abs/2017arXiv170304627P},
-  	Archiveprefix = {arXiv},
-  	Author = {{Pearson}, S. and {Price-Whelan}, A.~M. and {Johnston}, K.~V.},
-  	Eprint = {1703.04627},
-  	Journal = {ArXiv e-prints},
-  	Keywords = {Astrophysics - Astrophysics of Galaxies},
-  	Month = mar,
-    doi = {10.1038/s41550-017-0220-3},
-  	Title = {{Gaps in Globular Cluster Streams: Pal 5 and the Galactic Bar}},
-  	Year = 2017
+  url = {http://adsabs.harvard.edu/abs/2017arXiv170304627P},
+  Archiveprefix = {arXiv},
+  Author = {{Pearson}, S. and {Price-Whelan}, A.~M. and {Johnston}, K.~V.},
+  Eprint = {1703.04627},
+  Journal = {ArXiv e-prints},
+  Keywords = {Astrophysics - Astrophysics of Galaxies},
+  Month = mar,
+  doi = {10.1038/s41550-017-0220-3},
+  Title = {{Gaps in Globular Cluster Streams: Pal 5 and the Galactic Bar}},
+  Year = 2017
 }
 
 @article{azizi_minimax,
-  	Author = {{Azizi}, M. and {Wilhelm}, J. and {Panadés-Barrueta}, R.~L. and {Golze}, D. and {Giantomassi}, M. and {Delesma}, F.~A. and {Rinke}, P. and {Gonze}, X.},
-  	Journal = {Unpublished},
-    year = {2023},
-    doi = {--},
-  	Title = {{Validation of the Green-X library time-frequency component for GW and RPA calculations}},
+  Author = {{Azizi}, M. and {Wilhelm}, J. and {Panadés-Barrueta}, R.~L. and {Golze}, D. and {Giantomassi}, M. and {Delesma}, F.~A. and {Rinke}, P. and {Gonze}, X.},
+  Journal = {Unpublished},
+  year = {2023},
+  doi = {--},
+  Title = {{Validation of the Green-X library time-frequency component for GW and RPA calculations}},
 }
 @book{Binney:2008,
-  	url = {http://adsabs.harvard.edu/abs/2008gady.book.....B},
-  	Author = {{Binney}, J. and {Tremaine}, S.},
-  	Booktitle = {Galactic Dynamics: Second Edition, by James Binney and Scott Tremaine.~ISBN 978-0-691-13026-2 (HB).~Published by Princeton University Press, Princeton, NJ USA, 2008.},
[…]
