+++ docs/_config.yml
@@ -1,1 +1,1 @@
-theme: hacker
+theme: jekyll-theme-hacker
