+++ docs/index.md
@@ -5,13 +5,13 @@ tagline: GreenX
 description: Library for Many-body Greens Functions on HPC
 ---
 
-# GreenX
+# Overview
 
 A new open-source exascale library of Green-function-based methodologies. Its layered design will separate higher-level 
 functionalities (distinguishing between code-independent and code-family-specific parts) from architecture-dependent 
 numerical routines, common to all code families.
 
-## Components
+# Components
 
 - [Minimax Time-Frequency](gx_time_frequency.md)
 
