+++ docs/_config.yml
@@ -1,3 +1,6 @@
 remote_theme: pages-themes/hacker@v0.2.0
 plugins:
 - jekyll-remote-theme
+
+title: The GreenX library
+description: ""
