+++ docs/gx_ac.md
@@ -1,0 +1,18 @@
+---
+layout: default
+title: Analytic Continuation Component
+tagline: Analytic Continuation
+description: Analytic continuation routines for GW calculations
+---
+
+# General
+
+This component of the library is currently being developed, stay tuned!
+
+<button onclick="goBack()">Go Back</button>
+
+<script>
+function goBack() {
+  window.history.back();
+}
+</script>
