### 2023-08-02 #5570 @Panadestein
Dear @lucydot @mailhexu @DarioALeonValido,

thank you very much for your throughout review and valuable comments. We continued working on the remaining comments (although several of the raised issued were already solved, as @AlexBuccheri said). Now we have set up a GitHub pages website for the project ([nomad-coe.github.io/greenX/](https://nomad-coe.github.io/greenX/)), which hopefully solves the documentation comment. The website includes benchmark results and detail usage instructions, including a mapping from the variables in the paper to the actual variables in the code.

We hope that with these adjustments, our paper now meets all requirements for acceptance. Thank you for your thoughtful review and continued guidance.
