### 2023-08-07 8a9ba42
improve CH4 benchmark section
