### 2023-08-07 c48a60b
fix of latex syntax (website)
