+++ docs/index.md
@@ -7,9 +7,7 @@ description: Library for Many-body Greens Functions on HPC
 
 # Overview
 
-A new open-source exascale library of Green-function-based methodologies. Its layered design will separate higher-level 
-functionalities (distinguishing between code-independent and code-family-specific parts) from architecture-dependent 
-numerical routines, common to all code families.
+A new open-source library that supports exascale implementations of Green-function-based methodologies. Its layered design will separate higher-level functionalities (distinguishing between code-independent and code-family-specific parts) from architecture-dependent numerical routines, common to all code families.
 
 # Components
 
