+++ JOSS/paper.md
@@ -1,5 +1,5 @@
 ---
-title: ' Time-frequency component of the Green-X library: minimax grids for efficient RPA and \textit{GW} calculations.'
+title: ' Time-frequency component of the GreenX library: minimax grids for efficient RPA and \textit{GW} calculations.'
 tags:
   - FORTRAN
   - Low scaling \textit{GW} calculations
@@ -59,19 +59,19 @@ bibliography: refs.bib
 
 # Summary
 
-The central objects in many-body electronic structure theory, such as the \textit{GW} method and the random-phase approximation (RPA), are defined in the complex frequency or time domain. We present here a component of the Green-X library, which provides grids and weights for imaginary time-frequency transformations needed for Green's function based objects. The Green-X library emerged from the NOMAD Center of Excellence, whose objective is to enable accurate Green's function based electronic structure theory calculations on state-of-the-art supercomputers.
+The central objects in many-body electronic structure theory, such as the \textit{GW} method and the random-phase approximation (RPA), are defined in the complex frequency or time domain. We present here the GX-TimeFrequency component of the GreenX library, which provides grids and weights for imaginary time-frequency transformations needed for Green's function based objects. The GreenX library emerged from the NOMAD Center of Excellence, whose objective is to enable accurate Green's function based electronic structure theory calculations on state-of-the-art supercomputers.
 
-The package comprises minimax time and frequency grids [@Takatsuka2008; @kaltak2014low; @liu2016cubic] and corresponding quadrature weights to numerically compute time and frequency integrals of the correlation energy as well as weights for Fourier transforms between time and frequency grids. While we target low-scaling RPA and \textit{GW} algorithms, its compact frequency grids can also be used to reduce the computational prefactor in RPA implementations with conventional scaling. In addition, the time grids can be employed in Laplace-transformed direct MP2 (LT-dMP2) calculations. The Green-X source code is freely available on GitHub, and comes equipped with a build system, a comprehensive set of tests, and detailed documentation.
+The package comprises minimax time and frequency grids [@Takatsuka2008; @kaltak2014low; @liu2016cubic] and corresponding quadrature weights to numerically compute time and frequency integrals of the correlation energy as well as weights for Fourier transforms between time and frequency grids. While we target low-scaling RPA and \textit{GW} algorithms, its compact frequency grids can also be used to reduce the computational prefactor in RPA implementations with conventional scaling. In addition, the time grids can be employed in Laplace-transformed direct MP2 (LT-dMP2) calculations. The GreenX source code is freely available on GitHub, and comes equipped with a build system, a comprehensive set of tests, and detailed documentation.
 
 # Statement of need
 
[…]
