+++ GX-TimeFrequency/utilities/gx_tabulate_minimax.F90
@@ -1,5 +1,5 @@
 ! **************************************************************************************************
-!  Copyright (C) 2020-2023 Green-X library
+!  Copyright (C) 2020-2023 GreenX library
 !  This file is distributed under the terms of the APACHE2 License.
 !
 ! **************************************************************************************************
