+++ GX-TimeFrequency/test/test_gx_minimax_grid.py
@@ -1,6 +1,6 @@
 """
 ! **************************************************************************************************
-!  Copyright (C) 2020-2023 Green-X library
+!  Copyright (C) 2020-2023 GreenX library
 !  This file is distributed under the terms of the APACHE2 License.
 !
 ! **************************************************************************************************
