+++ GX-q=0/README.md
@@ -1,4 +1,4 @@
 ## GreenX library - q=0 
 
 This directory is a placeholder for the development of the q=0 (singularity treatment) component of GreenX.
-See the [GreenX document](Documents/Green-X.md) for the structure of the library. 
+See the [GreenX document](Documents/GreenX.md) for the structure of the library. 
