+++ GX-LocalizedBasis/README.md
@@ -1,4 +1,4 @@
 ## GreenX library - LocalizedBasis 
 
 This directory is a placeholder for the development of the LocalizedBasis component of GreenX.
-See the [GreenX document](Documents/Green-X.md) for the structure of the library. 
+See the [GreenX document](Documents/GreenX.md) for the structure of the library. 
