### 2023-08-16 #5570 @dgolze
> Dear @aziziph @AlexBuccheri @Panadestein -
>
> We are getting there - I have a few outstanding points:
>
>     * In your repo README you have the doxygen documentation listed but there is no link to the Github pages site; I think this should be added. I know there is a link at the top of the page but this can be missed when half way down, and I'd expect all docs to be listed under documentation section.
>
>     * Do you plan to update both forms of docs? This seems it might get a bit difficult to maintain and that there might be divergence. Users need to be clear about what is the master docs.
>
>     * The GH pages docs refer to an `exascale library`: perhaps wording such as `exascale-ready library` or `exascale-optimised` might be better - as the library itself is not exascale.
>
>     * Returning to a point made at the start of the review process: currently the paper is too long and goes into too deep a level of mathematical detail. I suggest you move the mathematical section to the online documentation. Even with this section removed from the paper you will pushing the allowed word limit I think. Detailed docs are meant to be hosted online, with the paper acting as a short 1/2-page advert for the code.
>
>     * I noticed slight inconsistencies with terminology. In the section `required input and output of the library` you make several references to a library when I think you are referring to the time-frequency component specifically. If so, this needs re-wording so that there is no confusion between the two.
>
>
> Happy to discuss more here, please let me know if anything is unclear, this is very much as a two-way process.
>
> Best,
>
> Lucy

Dear @lucydot,
cc @aziziph @Panadestein @AlexBuccheri @fdelesma

We have addressed points 1-3 and 5. A short additional comment about point 2: we plan to update both forms and also want to point out that the information in the README and the website doesn't overlap. The README contains instruction how to compile etc. and further information for developers, while the website covers usage examples, benchmarks etc. We removed remaining redundancies. About point  4, we have to wait for the input of the project's PI @gonzex.
Best,
Dorothea
