+++ README.md
@@ -97,7 +97,7 @@ Documentation is built in `documentation` and can be viewed by opening
 When adding new files with documentation, please ensure the directory is listed 
 in the `INPUT` tag of Doxyfile.
 
-For more information and benchmark examples see also the [GreenX website](nhttps://nomad-coe.github.io/greenX/).
+For more information and benchmark examples see also the [GreenX website](https://nomad-coe.github.io/greenX/).
 
 ## Unit Testing
 
