### 2023-08-18 291c91c
Removing the structure section to the website.
