+++ docs/index.md
@@ -7,9 +7,8 @@ description: Library for Many-body Greens Functions on HPC
 
 # Overview
 
-A new open-source library that supports exascale implementations of Green-function-based methodologies. Its layered design will separate higher-level functionalities (distinguishing between code-independent and code-family-specific parts) from architecture-dependent numerical routines, common to all code families.
-
-# Components
+A new open-source library that supports exascale implementations of Green-function-based methodologies. Its layered design will separate higher-level functionalities (distinguishing between code-independent and code-family-specific parts) from architecture-dependent numerical routines, common to all code families. The GreenX library will eventually provide a variety of tools for advanced electronic structure calculations. 
 - [Structure of the library](structure.md)
+# Components
 - [Minimax Time-Frequency](gx_time_frequency.md)
 - [Analytic Continuation](gx_ac.md)
