### 2023-08-19 310eeaa
moving description of the library from structure.md to the overview.
