### 2023-08-21 c88ebd9
shorten the beginning of the math part a bit
