### 2023-08-21 ed6f520
delete 7 references which are less relevant to the topic
