### 2023-08-21 7428811
put two equation on one line
