### 2023-08-21 8bd7cfe
fix label of equation
