+++ JOSS/paper.md
@@ -73,7 +73,7 @@ An important consideration for low-scaling algorithms is the crossover point. Du
 
 In this work, we present the GX-TimeFrequency component of the GreenX library, an open-source package distributed under the Apache license (Version 2.0). GX-TimeFrequency provides time and frequency grids and the corresponding integration weights to compute correlation energies for Green's function implementations. It also provides Fourier weights to convert between imaginary time and imaginary frequency. The library can be used for low-scaling RPA and \textit{GW} implementations. BSE codes, which use (low-scaling) \textit{GW} as input, can also utilize our library. The minimax grids are also suitable for RPA implementations with conventional scaling [@delben2015enabling]: The minimax grids are more compact than, e.g., Gauss-Legendre grids, resulting in a reduction of the computational prefactor, while yielding the same accuracy [@delben2015enabling;@azizi_minimax]. We note that the minimax grids are not recommended for conventional imaginary-frequency-only \textit{GW} implementations [@ren2012resolution;@wilhelm2016gw] since they have not been optimized for the frequency integral of the self-energy.
 
-While not being the main target of the library, the minimax time grids can also be utilized to calculate the LT-dMP2 correlation energy [@almloef1991elimination; @jung2004scaled; @Takatsuka2008; @kaltak2014low; @glasbrenner2020efficient]. The dMP2 term is one of two terms of the MP2 correlation energy. In a diagrammatic representation, dMP2 is the lowest order of the RPA correlation energy [@ren2012random]. The dMP2 correlation energy can be reformulated using the Laplace transform to obtain the LT-dMP2 expression which scales cubically in contrast to the $O(N^5)$ scaling of standard MP2.
+While not being the main target of the library, the minimax time grids can also be utilized to calculate the LT-dMP2 correlation energy [@almloef1991elimination; @jung2004scaled; @Takatsuka2008; @kaltak2014low; @glasbrenner2020efficient]. The dMP2 term is one of two terms of the MP2 correlation energy and, in a diagrammatic representation, corresponds to the lowest order of the RPA correlation energy [@ren2012random]. The dMP2 correlation energy can be reformulated using the Laplace transform to obtain the LT-dMP2 expression which scales cubically in contrast to the $O(N^5)$ scaling of standard MP2.
 
 # Mathematical framework
 
@@ -94,7 +94,7 @@ To convert between imaginary time and frequency grids, we introduce a nonuniform
 {F}(i\omega) =  {F}^\text{even}(i\omega) +  {F}^\text{odd}(i\omega)\,,\label{Fw_split}
 \end{align}
 with
-\begin{align}
+\begin{align}/
 \hat{F}^\text{even}(i\tau)=\hat{F}^\text{even}(-i\tau) \hspace{2.4em} &\text{and} \hspace{1.9em} F^\text{even}(i\omega)=F^\text{even}(-i\omega)\,, \\
 \hat{F}^\text{odd}(i\tau)=-\hat{F}^\text{odd}(-i\tau) \hspace{2em} &\text{and} \hspace{2em} F^\text{odd}(i\omega)=-F^\text{odd}(-i\omega)\,.
 \end{align}
@@ -111,7 +111,7 @@ The corresponding discrete Fourier transforms read [@liu2016cubic]
 \end{align}
[…]
