+++ JOSS/refs.bib
@@ -31,7 +31,7 @@
 @article{gaia,
   author = {{Gaia Collaboration}},
   title = "{The Gaia mission}",
-  journal = {Astronomy and Astrophysics},
+  journal = {Astron. Astrophys.},
   archivePrefix = "arXiv",
   eprint = {1609.04153},
   primaryClass = "astro-ph.IM",
@@ -46,7 +46,7 @@
 @article{astropy,
   author = {{Astropy Collaboration}},
   title = "{Astropy: A community Python package for astronomy}",
-  journal = {Astronomy and Astrophysics},
+  journal = {Astron. Astrophys.},
   archivePrefix = "arXiv",
   eprint = {1307.6212},
   primaryClass = "astro-ph.IM",
@@ -79,7 +79,7 @@
 @article{golze2019gw,
   title={The {GW} compendium: A practical guide to theoretical photoemission spectroscopy},
   author={Golze, Dorothea and Dvorak, Marc and Rinke, Patrick},
-  journal={Frontiers in chemistry},
+  journal={Front. Chem. },
   volume={7},
   pages={377},
   year={2019},
@@ -89,7 +89,7 @@
 @article{bruneval2008accurate,
   title={Accurate {GW} self-energies in a plane-wave basis using only a few empty states: Towards large systems},
   author={Bruneval, Fabien and Gonze, Xavier},
-  journal={Physical Review B},
+  journal={Phys. Rev. B},
   volume={78},
   number={8},
   pages={085125},
@@ -107,7 +107,7 @@
 @article{hackbusch2019computation,
   title={{Computation of best L$^\infty$ exponential sums for 1/x by Remez’ algorithm}},
   author={Hackbusch, Wolfgang},
-  journal={Computing and Visualization in Science},
+  journal={Comput. Vis. Sci.},
   volume={20},
   number={1-2},
   pages={1--11},
@@ -118,7 +118,7 @@
 @article{Takatsuka2008,
   author = {Takatsuka, Akio and Ten-no, Seiichiro and Hackbusch, Wolfgang},
   title = "{Minimax approximation for the decomposition of energy denominators in Laplace-transformed Møller–Plesset perturbation theories}",
-  journal = {The Journal of Chemical Physics},
+  journal = {J. Chem. Phys.},
   volume = {129},
   number = {4},
   year = {2008},
@@ -128,7 +128,7 @@
 @article{Adler1962,
   author = "Adler, Stephen L.",
   issn = "0031-899X",
-  journal = "Physical Review",
+  journal = "Phys. Rev.",
   month = "April",
   number = "2",
   pages = "413-420",
@@ -142,7 +142,7 @@
 @article{Wiser1963,
   author = "Wiser, Nathan",
   issn = "0031-899X",
-  journal = "Physical Review",
+  journal = "Phys. Rev.",
   month = "January",
   number = "1",
   pages = "62-69",
@@ -156,7 +156,7 @@
 @article{haser1992laplace,
   title={Laplace transform techniques in Mo/ller-Plesset perturbation theory},
   author={H{\"a}ser, Marco and Alml{\"o}f, Jan},
-  journal={The Journal of Chemical Physics},
+  journal={J. Chem. Phys.},
   volume={96},
   number={1},
   pages={489--494},
@@ -167,7 +167,7 @@
 @article{kaltak2014low,
   title={Low scaling algorithms for the random phase approximation: Imaginary time and Laplace transformations},
   author={Kaltak, Merzuk and Klime{\v{s}}, Ji{\v{r}}\'{i} and Kresse, Georg},
-  journal={Journal of Chemical Theory and Computation},
+  journal={J. Chem. Theory Comput. },
   volume={10},
   number={6},
   pages={2498--2507},
@@ -178,7 +178,7 @@
[…]
