+++ JOSS/refs.bib
@@ -211,7 +211,7 @@
 @article{rojas1995space,
   title={Space-time method for ab initio calculations of self-energies and dielectric response functions of solids},
   author={Rojas, HN and Godby, Rex William and Needs, RJ},
-  journal={Phys. Rev. Letters},
+  journal={Phys. Rev. Lett},
   volume={74},
   number={10},
   pages={1827},
@@ -538,14 +538,15 @@
   YEAR={2021},
 }
 @article{foerster2023twocomponent,
-  title={{Two-component GW calculations: Cubic scaling implementation and comparison of partially self-consistent variants}},
-  author={Arno Förster and Erik van Lenthe and Edoardo Spadetto and Lucas Visscher},
-  year={2023},
-  pages={2303.09979},
-  doi = {10.48550/arXiv.2303.09979},
-  journal={arXiv},
+ author  = {Förster, Arno and van Lenthe, Erik and Spadetto, Edoardo and Visscher, Lucas},
+  journal = {J. Chem. Theory Comput.},
+  title   = {Two-Component GW Calculations: Cubic Scaling Implementation and Comparison of Vertex-Corrected and Partially Self-Consistent GW Variants},
+  year    = {2023},
+  doi     = {10.1021/acs.jctc.3c00512},
+  eprint  = { https://doi.org/10.1021/acs.jctc.3c00512 },
+  url     = {https://doi.org/10.1021/acs.jctc.3c00512
+},
 }
-
 @article{glasbrenner2020efficient,
   author = {Glasbrenner, Michael and Graf, Daniel and Ochsenfeld, Christian},
   title = {{Efficient Reduced-Scaling Second-Order Møller–Plesset Perturbation Theory with Cholesky-Decomposed Densities and an Attenuated Coulomb Metric}},
