### 2023-08-22 958eca5
some other small fixes of bugs introduced while shortening
