+++ JOSS/refs.bib
@@ -543,9 +543,7 @@
   title   = {Two-Component GW Calculations: Cubic Scaling Implementation and Comparison of Vertex-Corrected and Partially Self-Consistent GW Variants},
   year    = {2023},
   doi     = {10.1021/acs.jctc.3c00512},
-  eprint  = { https://doi.org/10.1021/acs.jctc.3c00512 },
-  url     = {https://doi.org/10.1021/acs.jctc.3c00512
-},
+  url     = {https://doi.org/10.1021/acs.jctc.3c00512},
 }
 @article{glasbrenner2020efficient,
   author = {Glasbrenner, Michael and Graf, Daniel and Ochsenfeld, Christian},
