+++ docs/structure.md
@@ -46,7 +46,8 @@ In this section, we focus on the components of the library which are ready to us
   |- README.md
 ```
 
-The time and frequency grids $$\{\tau_k(R)\}_{k=1}^n$$, $$\{\omega_j(R)\}_{j=1}^n$$ and the integration weights $$\{\gamma_k(R)\}_{k=1}^n$$, $$\{\sigma_j(R)\}_{j=1}^n$$ are tabulated (hard coded) in the module 'minimax_tau.f90' and 'minimax_omega.f90' for various ranges $$R$$ of interest for solids and molecules. The module 'minimax_grids.f90' calculates the weights $$\{\delta_{kj}\}_{k,j=1}^n$$, $$\{\eta_{jk}\}_{k,j=1}^n$$, $$\{\lambda_{kj}\}_{k,j=1}^n$$ of cosine and sine transforms during the runtime of the program. 'minimax_grids.f90' also provides the duality error $$\Delta_{\text{CT}}$$. 'minimax_utils.f90' contains auxiliary procedures and data structures for the main minimax routines.
+The time and frequency grids $$\{\tau_k(R)\}_{k=1}^n$$, $$\{\omega_j(R)\}_{j=1}^n$$ and the integration weights $$\{\gamma_k(R)\}_{k=1}^n$$, $$\{\sigma_j(R)\}_{j=1}^n$$ are tabulated (hard coded) in the module `minimax_tau.f90` and `minimax_omega.f90` for various ranges $$R$$ of interest for solids and molecules. The module `minimax_grids.f90` calculates the weights $$\{\delta_{kj}\}_{k,j=1}^n$$, $$\{\eta_{jk}\}_{k,j=1}^n$$, $$\{\lambda_{kj}\}_{k,j=1}^n$$ of cosine and sine transforms during the runtime of the program. 'minimax_grids.f90' also provides the duality error $$\Delta_{\text{CT}}$$. `minimax_utils.f90` contains auxiliary procedures and data structures for the main minimax routines.
+
 <button onclick="goBack()">Go Back</button>
 
 <script>
