### 2023-08-22 0559b1f
Remove superfluous headers
