### 2023-08-22 87f428d
Further paper shortening
