+++ JOSS/refs.bib
@@ -115,6 +115,16 @@
   doi = {10.1007/s00791-018-00308-4},
   publisher={Springer}
 }
+@article{panadesbarrueta2023,
+author = {Panadés-Barrueta, Ramón L. and Golze, Dorothea},
+title = {Accelerating Core-Level GW Calculations by Combining the Contour Deformation Approach with the Analytic Continuation of W},
+journal = {J. Chem. Theory Comput.},
+volume = {19},
+number = {16},
+pages = {5450-5464},
+year = {2023},
+doi = {10.1021/acs.jctc.3c00555},
+}
 @article{Takatsuka2008,
   author = {Takatsuka, Akio and Ten-no, Seiichiro and Hackbusch, Wolfgang},
   title = "{Minimax approximation for the decomposition of energy denominators in Laplace-transformed Møller–Plesset perturbation theories}",
