### 2023-08-23 4673eb2
Eq and citation
