+++ JOSS/refs.bib
@@ -288,6 +288,13 @@ doi = {10.1021/acs.jctc.3c00555},
   doi = {10.1103/PhysRevB.71.193102},
   publisher={APS}
 }
+@Article{Shi2023,
+  author  = {Rong Shi and Peize Lin and Min-Ye Zhang and Lixin He and Xinguo Ren},
+  journal = {arXiv:2307.12029},
+  title   = {Sub-quadratic scaling real-space random-phase approximation correlation energy calculations for periodic systems with numerical atomic orbitals},
+  year    = {2023},
+  url     = {https://arxiv.org/abs/2307.12029},
+}
 @article{wilhelm2016gw,
   title={{GW} in the {G}aussian and plane waves scheme with application to linear acenes},
   author={Wilhelm, Jan and Del Ben, Mauro and Hutter, J\"{u}rg},
