### 2023-08-23 af06d24
Improve spacing in Eq. (2)-(5)
