### 2023-08-23 582705f
Correct "Gaussian" instead of "gaussian" in reference (btw, the upper/lower cases in journal titles are often incorrect)
