+++ JOSS/refs.bib
@@ -716,20 +716,9 @@ and Furche, Filipp},
   year={2015},
   publisher={APS}
 }
-@article{wilhelm2016large,
-  title={Large-scale cubic-scaling random phase approximation correlation energy calculations using a Gaussian basis},
-  author={Wilhelm, Jan and Seewald, Patrick and Del Ben, Mauro and Hutter, Jurg},
-  journal={J. Chem. Theory Comput. },
-  volume={12},
-  number={12},
-  pages={5851--5859},
-  year={2016},
-  doi = {10.1021/acs.jctc.6b00840},
-  publisher={ACS Publications}
-}
 @article{wilhelm2016rpa,
   author = {Wilhelm, Jan and Seewald, Patrick and Del Ben, Mauro and Hutter, Jürg},
-  title = {Large-Scale Cubic-Scaling Random Phase Approximation Correlation Energy Calculations Using a Gaussian Basis},
+  title={{Large-scale cubic-scaling random phase approximation correlation energy calculations using a Gaussian basis}},
   journal = {J. Chem. Theory Comput. },
   volume = {12},
   number = {12},
