### 2023-08-23 f67f5d4
Improve spacing in Eq. (2)-(5)
