### 2023-08-25 4902190
Update paper.md
Remove unneeded words.
