### 2023-08-25 93cb571
Update paper.md
More removal of unneeded words.
