### 2023-08-28 #5570 @gonzex
Dear Lucy,

Thanks for your comments on our manuscript.
We have revised it, and we hope it is now acceptable for publication in JOSS.
In view of the paper to be more “advert”-like, we agree that the length was a concern. We have thus carefully examined the whole content of the paper. The section “Structure of the library” did not fit in this picture, and we have moved it to the Web page of GreenX.  The “Mathematical framework” section has been reduced by half : we only have kept the minimum needed for the section “Required Input and Output” to be understandable by the reader. The latter is not a detailed description of the input/output, but it allows the reader to understand what the code is doing.  We think that the people potentially interested need such information even in an advert-type publication. We then also performed miscellaneous removals inside the manuscript. Altogether, the paper has been reduced from 10.5 pages to 7 pages, among those 2.5 pages being bibliography (that we think is just fair to keep).
This is not an exceptional length for a JOSS paper. In order to better realize whether we are in an acceptable range, we surveyed the 40 last published papers in JOSS. On Friday 25 August, joss.04913 was published. It is a 7 pages, and also includes 2.5 pages of bibliography. Thus its size is similar to ours. In the 40 last published papers, five of them are 7 pages or more (joss.04913, joss.05305,joss.05319, joss.05076, joss.05466). The longest is joss.05466 with 14 pages. We agree that we are still on the larger side, but such length is apparently acceptable.
May we thus respectfully ask you to accept our manuscript in the present form ?

With our best regards,
The authors
