### 2023-09-05 37e0476
introduce FF_BASE for common Fortran flags
The original STD_FFLAGS are only for GCC.
