### 2023-09-05 5d0b4fe
Deprecate -fp in Intel debug flags
This flag is no longer supported after 2020 update 4.
