### 2023-09-05 #5570 @lucydot
Thanks for the update and paper edits @gonzex and team.

I'm going to discuss this with the wider JOSS editorial team, as I think you have highlighted an inconsistency between what we ask for, and what is being published. I hope to get back to you in a week with this.
