### 2023-09-05 #5570 @lucydot
It seems that papers < 10 pages in length are somewhat of a grey area. I'm happy to accept the paper at its current length, especially given that effort has already been put into editing down.
