### 2023-09-05 #5570 @lucydot
Dear Green-X team (@aziziph @Panadestein @AlexBuccheri @fdelesma @dgolze)

- There is an invalid DOI ("--") that needs to be updated/removed - this is for the unpublished work cited.
- I'll generate the post-review checklist. You can't tick off the items, but please let me know when each item has been completed here in the thread.
