+++ JOSS/refs.bib
@@ -11,13 +11,6 @@
   Year = 2017
 }
 
-@article{azizi_minimax,
-  Author = {{Azizi}, M. and {Wilhelm}, J. and {Panadés-Barrueta}, R.~L. and {Golze}, D. and {Giantomassi}, M. and {Delesma}, F.~A. and {Rinke}, P. and {Gonze}, X.},
-  Journal = {Unpublished},
-  year = {2023},
-  doi = {--},
-  Title = {{Validation of the GreenX library time-frequency component for GW and RPA calculations}},
-}
 @book{Binney:2008,
   url = {http://adsabs.harvard.edu/abs/2008gady.book.....B},
   Author = {{Binney}, J. and {Tremaine}, S.},
