### 2023-09-06 #5570 @aziziph
Dear @lucydot

We would like to thank you for the acceptance. Subsequently, we have completed the "Author Tasks After Review is Complete".

- Double check authors and affiliations (including ORCIDs) DONE!
- Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper. DONE!
- Archive the release on Zenodo/figshare/etc and post the DOI here. DONE!
doi:10.5281/zenodo.8321618
- Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper. DON!
- Make sure that the license listed for the archive is the same as the software license. DONE!

Best regards,
On behalf of the authors
