### 2023-09-07 0b42ff4
Remove tilde; correct: 34 minimax points instead of 24
