### 2023-09-07 6d46884
correct the comments, remove commented -fp option
