### 2023-09-11 39d22f4
Remove "_" in labels trying to fix generated XML metadata file
