### 2023-09-13 #5570 @aziziph
Dear @lucydot,

The editorialbot encountered some issues during the paper acceptance process, primarily stemming from markdown problems. We have addressed these issues, and it appears that they have now been successfully resolved. We kindly request you to please rerun the "recommend-accept" function to confirm its functionality.

We apologize for any inconvenience this may have caused and sincerely appreciate your cooperation. Thank you in advance!

Best regards,
Maryam
