### 2023-09-16 74b1f75
Extend the CI to test with latest Intel one-API
