### 2023-09-18 #5570 @dgolze
Dear @lucydot,
can you run the recommend-accept command again please? As far as we know, only editors can do this. We checked and fixed the warnings of the logs in the GitHub Action. We hope everything works now.

Best regards,
Dorothea
