### 2023-09-22 e14433b
Run the oneAPI tests
