+++ .github/workflows/action.yml
@@ -118,3 +118,13 @@ jobs:
           cd build
           cmake -DENABLE_UNITTESTS=ON -DZOFU_PATH=/home/runner/work/greenX/greenX/external/zofu/install ../
           make -j$(nproc) VERBOSE=1
+
+      - name: Run Tests
+        run: |
+          cd python && pip install -e . && cd ../
+          cd build
+          ctest -j1 --output-on-failure
+        # The build system exports GX_BUILD_DIR, but I do not know how to
+        # propagate env vars between steps in Github Actions
+        env:
+          GX_BUILD_DIR: "/home/runner/work/greenX/greenX/build"
