### 2023-09-22 87107fb
Add ifort/ifx matrix, re-enable build test
