### 2023-09-25 #5570 @Panadestein
Hello @arfon, thanks for the recommend-accept. We have been facing this XML error for a while now. I looked in the documentation of JOSS, but did not find any entry about how to deal with it. Do you have any hint?  Thanks a lot
