### 2023-09-25 0dda8af
replacing align with eqnarray
