### 2023-09-25 #5570 @arfon
I think the issue here is that Pandoc is struggling to identify the equations labeled/linked in your document. Have you tried following these docs? https://joss.readthedocs.io/en/latest/submitting.html#equations
