### 2023-09-25 #5570 @aziziph
Dear @arfon,

We have updated the md file with the labeling format explained in the documentation. Could you please run the recommend-accept again to see if it works.

Thank you very much in advance!
Best,
Maryam
