### 2023-09-27 #5570 @lucydot
@aziziph @dgolze apologies for my late reply to your earlier messages.

It seems there is still a problem @arfon, will raise with other editors in case they have seen similar.
