### 2023-09-27 #5570 @dgolze
> Is the `hspace` necessary @aziziph? As a test could we try removing it to see if that is where the problem is? Just to check, is the `hspace` being used in math mode? I don't think valid latex if so.@lucydot
we will remove the hspace, trying right now
