### 2023-09-28 079adde
Added logo to README.md
