### 2023-09-28 4cb83d5
Reduce size of logo
