### 2023-10-03 b1602bc
\Delta_{CT} is replaced by \Delta_{FB}
