### 2023-10-03 738416c
replace duality error to forward backward error
