### 2023-10-03 #5570 @kyleniemeyer
@dgolze just a question: can you explain why GW is italicized in the title and text, while other acronyms like RPA are not?
