### 2023-10-03 #5570 @kyleniemeyer
@openjournals/dev looks like some LaTeX markup got caught in the title at https://doi.org/10.21105/joss.05570, does that need to be fixed manually?
