### 2023-10-07 #5570 @Panadestein
@kyleniemeyer the article already got indexed by Google Scholar and ResearchGate with this latex code in the title. That does not look good for anybody. Would you mind taking a look at this issue? Thanks
