### 2023-10-08 #5570 @arfon
@aziziph – please merge this PR. @kyleniemeyer then you can run reaccept here: https://joss.readthedocs.io/en/latest/editorial_bot.html?highlight=reaccept#updating-an-already-accepted-paper
