### 2023-10-09 #5570 @kyleniemeyer
Thanks, @arfon. I thought fixing this would involve some manual magic on the website.

@Panadestein @dgolze the website appears correct for me now, and indexers should update after some time as well.
